+++ app/build.gradle.kts
@@ -44,6 +44,7 @@ android {
 
 dependencies {
     implementation(project(":domain"))
+    implementation(project(":di"))
 
     implementation("androidx.core:core-ktx:1.10.1")
     implementation("androidx.appcompat:appcompat:1.6.1")
@@ -51,6 +52,7 @@ dependencies {
     implementation("androidx.constraintlayout:constraintlayout:2.1.4")
     implementation("androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.1")
     implementation("androidx.lifecycle:lifecycle-extensions:2.2.0")
+    implementation(project(mapOf("path" to ":di")))
     testImplementation("junit:junit:4.13.2")
     testImplementation("com.google.truth:truth:1.1.3")
     testImplementation("androidx.arch.core:core-testing:2.2.0")
@@ -70,5 +72,4 @@ dependencies {
     testImplementation("org.robolectric:robolectric:4.9")
 
     implementation("androidx.fragment:fragment-ktx:1.5.7")
-    implementation(kotlin("reflect"))
 }
