--- app/src/main/java/woowacourse/shopping/di/annotation/InMemory.kt
@@ -1,4 +1,0 @@
-package woowacourse.shopping.di.annotation
-
-@Qualifier
-annotation class InMemory
