+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -4,11 +4,11 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import io.hyemdooly.di.annotation.InDisk
+import io.hyemdooly.di.annotation.InMemory
 import kotlinx.coroutines.launch
 import woowacourse.shopping.data.CartRepository
 import woowacourse.shopping.data.ProductRepository
-import woowacourse.shopping.di.annotation.InDisk
-import woowacourse.shopping.di.annotation.InMemory
 import woowacourse.shopping.model.Product
 
 class MainViewModel constructor(
