+++ app/src/main/java/woowacourse/shopping/ui/ViewModelFactory.kt
@@ -1,8 +1,9 @@
-package woowacourse.shopping.di
+package woowacourse.shopping.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.CreationExtras
+import io.hyemdooly.di.Injector
 
 val viewModelFactory = object : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>, extras: CreationExtras): T =
