+++ di/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    kotlin("jvm")
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
+}
+dependencies {
+    implementation(kotlin("reflect"))
+}
