+++ di/src/main/java/io/hyemdooly/di/annotation/InDisk.kt
@@ -1,0 +1,4 @@
+package io.hyemdooly.di.annotation
+
+@Qualifier
+annotation class InDisk
