+++ di/src/main/java/io/hyemdooly/di/annotation/InMemory.kt
@@ -1,0 +1,4 @@
+package io.hyemdooly.di.annotation
+
+@Qualifier
+annotation class InMemory
