+++ di/src/main/java/io/hyemdooly/di/annotation/Inject.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di.annotation
+package io.hyemdooly.di.annotation
 
 @Target(AnnotationTarget.FIELD, AnnotationTarget.PROPERTY)
 annotation class Inject
