+++ di/src/main/java/io/hyemdooly/di/annotation/Qualifier.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di.annotation
+package io.hyemdooly.di.annotation
 
 @Target(AnnotationTarget.ANNOTATION_CLASS)
 annotation class Qualifier
